// src/Moryx.Resources.Web/app/src/app/api/models/resource-reference-model.ts
/*
 * Copyright (c) 2026 Phoenix Contact GmbH & Co. KG
 * Licensed under the Apache License, Version 2.0
*/

/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ResourceModel } from '../models/resource-model';
export interface ResourceReferenceModel {
  name?: string | null;
  targets?: Array<ResourceModel> | null;
}